Use an extension type for ide theme query parameters
